migrate to make use  of entry.runtime_data
