+++ homeassistant/components/shopping_list/intent.py
@@ -5,7 +5,8 @@ from __future__ import annotations
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import config_validation as cv, intent
 
-from . import DOMAIN, EVENT_SHOPPING_LIST_UPDATED, NoMatchingShoppingListItem
+from .common import NoMatchingShoppingListItem, _get_shopping_data
+from .const import DOMAIN, EVENT_SHOPPING_LIST_UPDATED
 
 INTENT_ADD_ITEM = "HassShoppingListAddItem"
 INTENT_COMPLETE_ITEM = "HassShoppingListCompleteItem"
@@ -31,7 +32,7 @@ class AddItemIntent(intent.IntentHandler):
         """Handle the intent."""
         slots = self.async_validate_slots(intent_obj.slots)
         item = slots["item"]["value"].strip()
-        await intent_obj.hass.data[DOMAIN].async_add(item)
+        await _get_shopping_data(intent_obj.hass).async_add(item)
 
         response = intent_obj.create_response()
         intent_obj.hass.bus.async_fire(EVENT_SHOPPING_LIST_UPDATED)
@@ -52,7 +53,9 @@ class CompleteItemIntent(intent.IntentHandler):
         item = slots["item"]["value"].strip()
 
         try:
-            complete_items = await intent_obj.hass.data[DOMAIN].async_complete(item)
+            complete_items = await _get_shopping_data(intent_obj.hass).async_complete(
+                item
+            )
         except NoMatchingShoppingListItem:
             complete_items = []
 
@@ -74,13 +77,13 @@ class ListTopItemsIntent(intent.IntentHandler):
 
     async def async_handle(self, intent_obj: intent.Intent) -> intent.IntentResponse:
         """Handle the intent."""
-        items = intent_obj.hass.data[DOMAIN].items[-5:]
+        items = _get_shopping_data(intent_obj.hass).items[-5:]
         response: intent.IntentResponse = intent_obj.create_response()
 
         if not items:
             response.async_set_speech("There are no items on your shopping list")
         else:
-            items_list = ", ".join(itm["name"] for itm in reversed(items))
+            items_list = ", ".join(str(itm["name"]) for itm in reversed(items))
             response.async_set_speech(
                 f"These are the top {min(len(items), 5)} items on your shopping list: {items_list}"
             )
